+++ src/cmd/script.rs
@@ -44,6 +44,11 @@ pub fn new<'a, 'b>() -> App<'a, 'b> {
                 .takes_value(true)
                 .multiple(true),
         )
+        .arg(
+            Arg::with_name("abort-on-error")
+                .long("abort-on-error")
+                .help("Abort analysis/compilation on first caught error")
+        )
 }
 
 /// Execute the `script` subcommand.
@@ -59,6 +64,8 @@ pub fn run(sess: &Session, matches: &ArgMatches) -> Result<()> {
         "vivado" => &["vivado", "synthesis", "fpga", "xilinx"],
         _ => unreachable!(),
     };
+
+    let abort_on_error = matches.is_present("abort-on-error");
 
     // Filter the sources by target.
     let targets = matches
@@ -81,8 +88,8 @@ pub fn run(sess: &Session, matches: &ArgMatches) -> Result<()> {
 
     // Generate the corresponding output.
     match matches.value_of("format").unwrap() {
-        "vsim" => emit_vsim_tcl(sess, matches, targets, srcs),
-        "synopsys" => emit_synopsys_tcl(sess, matches, targets, srcs),
+        "vsim" => emit_vsim_tcl(sess, matches, targets, srcs, abort_on_error),
+        "synopsys" => emit_synopsys_tcl(sess, matches, targets, srcs, abort_on_error),
         "vivado" => emit_vivado_tcl(sess, matches, targets, srcs),
         _ => unreachable!(),
     }
@@ -130,12 +137,22 @@ enum SourceType {
     Vhdl,
 }
 
+fn tcl_catch_prefix(cmd: &str, do_prefix: bool) -> String {
+    let prefix = if do_prefix { "if {[catch {" } else { "" };
+    return format!("{}{}", prefix, cmd);
+}
+
+fn tcl_catch_postfix() -> &'static str {
+    return "}]} {return 1}";
+}
+
 /// Emit a vsim compilation script.
 fn emit_vsim_tcl(
     sess: &Session,
     matches: &ArgMatches,
     targets: TargetSet,
     srcs: Vec<SourceGroup>,
+    abort_on_error: bool,
 ) -> Result<()> {
     println!("# This script was generated automatically by bender.");
     println!("set ROOT \"{}\"", sess.root.to_str().unwrap());
@@ -154,7 +171,7 @@ fn emit_vsim_tcl(
                 let mut lines = vec![];
                 match ty {
                     SourceType::Verilog => {
-                        lines.push("vlog -incr -sv".to_owned());
+                        lines.push(tcl_catch_prefix("vlog -incr -sv", abort_on_error).to_owned());
                         if let Some(args) = matches.values_of("vlog-arg") {
                             lines.extend(args.map(Into::into));
                         }
@@ -186,7 +203,7 @@ fn emit_vsim_tcl(
                         }
                     }
                     SourceType::Vhdl => {
-                        lines.push("vcom -2008".to_owned());
+                        lines.push(tcl_catch_prefix("vcom -2008", abort_on_error).to_owned());
                         if let Some(args) = matches.values_of("vcom-arg") {
                             lines.extend(args.map(Into::into));
                         }
@@ -208,6 +225,9 @@ fn emit_vsim_tcl(
                 }
                 println!("");
                 println!("{}", lines.join(" \\\n    "));
+                if abort_on_error {
+                    println!("{}", tcl_catch_postfix());
+                }
             },
         );
     }
@@ -220,6 +240,7 @@ fn emit_synopsys_tcl(
     _matches: &ArgMatches,
     targets: TargetSet,
     srcs: Vec<SourceGroup>,
+    abort_on_error: bool,
 ) -> Result<()> {
     println!("# This script was generated automatically by bender.");
     println!("set search_path_initial $search_path");
@@ -255,10 +276,10 @@ fn emit_synopsys_tcl(
             },
             |src, ty, files| {
                 let mut lines = vec![];
-                lines.push(format!("analyze -format {}", match ty {
+                lines.push(tcl_catch_prefix(&format!("analyze -format {}", match ty {
                         SourceType::Verilog => { "sv" }
                         SourceType::Vhdl => { "vhdl" }
-                    }).to_owned());
+                    }), abort_on_error).to_owned());
 
                 // Add defines.
                 let mut defines: Vec<(String, Option<&str>)> = vec![];
@@ -294,6 +315,9 @@ fn emit_synopsys_tcl(
                 lines.push("]".to_owned());
                 println!("");
                 println!("{}", lines.join(" \\\n    "));
+                if abort_on_error {
+                    println!("{}", tcl_catch_postfix());
+                }
             },
         );
     }